Update runtime.rst
Added note to account for `publickey` authentication issues on commands such as `gigalixir ps:migrate`, yet `git push` still succeeds. If this is the case, it is likely (though not guaranteed) that multiple SSH keys are present. The Gigalixir CLI might not auto-select the desired key when multiple keys are present, and the error message presented by the CLi isn't immediately intuitive.

The note provides some troubleshooting suggestions.