Pin docutils to a specific version
According to pipdeptree, it looks like sphinx-tabs wants ~=0.16.0 and
Sphinx wants >=0.14,<0.18. This means 0.16.0 should work, but somehow
Sphinx is choosing 0.17.0 which is incompatible with sphinx-tabs, at
least until 4.0 (currently on 3.2.0). I'm not sure why python's
dependency resolution isn't figuring this out, but pinning the version
manually like this seems to work.